src/main/java/it/schm/magnolia/events/ui/converter/ZoneIdConverter.java
/*
 * MIT License
 *
 * Copyright (c) 2020 Sam Schmit-Van Werweke
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

package it.schm.magnolia.events.ui.converter;

import com.vaadin.data.Converter;
import com.vaadin.data.Result;
import com.vaadin.data.ValueContext;

import javax.inject.Singleton;

import java.time.ZoneId;
import java.util.Objects;

/**
 * A {@link Converter} implementation capable of converting to and from {@link ZoneId}s.
 */
@Singleton
public class ZoneIdConverter implements Converter<ZoneId, String> {

    @Override
    public Result<String> convertToModel(ZoneId value, ValueContext context) {
        return Result.ok(Objects.isNull(value) ? null : value.getId());
    }

    @Override
    public ZoneId convertToPresentation(String value, ValueContext context) {
        return Objects.isNull(value) ? null : ZoneId.of(value);
    }

}
